# openapi.yaml
openapi: 3.0.3
info:
  title: Dogfood 2026 REST API
  version: 1.0.0
  description: Same-origin REST API. Protected operations use the dogfood_session httpOnly cookie.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /events/{slug}/teams:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    post:
      tags: [Teams and submissions]
      summary: Create a participant team
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [name], properties: { name: { type: string } } }
      responses:
        '201': { description: Team and invite token }
        '401': { $ref: '#/components/responses/Error' }
    put:
      tags: [Teams and submissions]
      summary: Join a team by invite token
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [inviteToken], properties: { inviteToken: { type: string } } }
      responses:
        '200': { description: Membership created }
        '409': { $ref: '#/components/responses/Error' }
...
  /events/{slug}/votes:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    post:
      tags: [Public participation]
      summary: Cast one community vote (participant)
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [submissionId], properties: { submissionId: { type: string } } }
      responses:
        '201': { description: Vote recorded }
        '409': { description: Duplicate vote or voting closed }
        '429': { description: Rate limited }
...
  /events/{slug}/results:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Judging]
      summary: View results; normalized scores selected with x-score-view header
      parameters: [{ name: x-score-view, in: header, schema: { type: string, enum: [raw, normalized], default: raw } }]
      responses:
        '200': { description: Published results or manager view }
        '403': { description: Results hidden }
    post:
      tags: [Judging]
      summary: Publish results after judging and voting end (owner/admin)
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Results published }, '409': { description: Judging or voting still active } }
...
  /events/{slug}/records:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Operations]
      summary: Public signed judge participation records
      responses: { '200': { description: Payload, SHA-256 hash, HMAC, and issue time } }
    post:
      tags: [Operations]
      summary: Issue/update participation records (owner/admin)
      security: [{ cookieAuth: [] }]
      responses: { '201': { description: Issued records } }
  /events/{slug}/webhooks:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Operations]
      summary: List webhook configuration and deliveries (owner/admin)
      security: [{ cookieAuth: [] }]
      responses: { '200': { description: Webhooks and recent attempts } }
    post:
      tags: [Operations]
      summary: Add signed webhook destination (owner/admin)
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [url], properties: { url: { type: string, format: uri }, enabled: { type: boolean } } }
      responses: { '201': { description: Webhook config and secret } }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dogfood_session
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: JSON error object
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }